controller/requirements.txt
fastapi
uvicorn[standard]
motor
cassandra-driver
python-dotenv
